test: add unit test for user schema method "createAuthToken"
